irrigation-intelligente/requirements.txt
pandas
numpy
scikit-learn
dvc
flask
pyyaml
joblib
pytest